Refactor
Исправлены разные приведения типов

Сейчас проблема при запуске fluttershy.cor и slider2.cor на 7549 цикле

+++ srcs/op_funcs_11_15.c
@@ -23,7 +23,7 @@ void			sti_func(t_rules *rules, t_champion *cursor, unsigned char *t_args)
 	arg2 = get_arg(rules, cursor, t_args, 1);
 	arg3 = get_arg(rules, cursor, t_args, 2);
 	if (t_args[1] == IND_CODE)
-		arg2 = (int)get_value_from_battlefield(rules, cursor->position,
+		arg2 = (short)get_value_from_battlefield(rules, cursor->position,
 										  arg2 % IDX_MOD, REG_SIZE);
 	else if (t_args[1] == REG_CODE)
 		arg2 = cursor->reg[arg2];
@@ -78,7 +78,7 @@ void			lld_func(t_rules *rules, t_champion *cursor, unsigned char *t_args)
 	arg2 = get_arg(rules, cursor, t_args, 1);
 
 	if (t_args[0] == IND_CODE)
-		arg1 = get_value_from_battlefield(rules, cursor->position,
+		arg1 = (short)get_value_from_battlefield(rules, cursor->position,
 										  arg1, REG_SIZE);
 	cursor->reg[arg2] = arg1;
 	cursor->carry = (cursor->reg[arg1] == 0) ? 1 : 0;
