+++ srcs/op_funcs_6_10.c
@@ -23,12 +23,12 @@ void	and_func(t_rules *rules, t_champion *cursor, unsigned char *t_args)
 	arg2 = get_arg(rules, cursor, t_args, 1);
 	arg3 = get_arg(rules, cursor, t_args, 2);
 	if (t_args[0] == IND_CODE)
-		arg1 = get_value_from_battlefield(rules, cursor->position,
+		arg1 = (short)get_value_from_battlefield(rules, cursor->position,
 				arg1 % IDX_MOD, REG_SIZE);
 	else if (t_args[0] == REG_CODE)
 		arg1 = cursor->reg[arg1];
 	if (t_args[1] == IND_CODE)
-		arg2 = get_value_from_battlefield(rules, cursor->position,
+		arg2 = (short)get_value_from_battlefield(rules, cursor->position,
 				arg2 % IDX_MOD, REG_SIZE);
 	else if (t_args[1] == REG_CODE)
 		arg2 = cursor->reg[arg2];
@@ -46,12 +46,12 @@ void			or_func(t_rules *rules, t_champion *cursor, unsigned char *t_args)
 	arg2 = get_arg(rules, cursor, t_args, 1);
 	arg3 = get_arg(rules, cursor, t_args, 2);
 	if (t_args[0] == IND_CODE)
-		arg1 = get_value_from_battlefield(rules, cursor->position,
+		arg1 = (short)get_value_from_battlefield(rules, cursor->position,
 										  arg1 % IDX_MOD, REG_SIZE);
 	else if (t_args[0] == REG_CODE)
 		arg1 = cursor->reg[arg1];
 	if (t_args[1] == IND_CODE)
-		arg2 = get_value_from_battlefield(rules, cursor->position,
+		arg2 = (short)get_value_from_battlefield(rules, cursor->position,
 										  arg2 % IDX_MOD, REG_SIZE);
 	else if (t_args[1] == REG_CODE)
 		arg2 = cursor->reg[arg2];
@@ -69,12 +69,12 @@ void			xor_func(t_rules *rules, t_champion *cursor, unsigned char *t_args)
 	arg2 = get_arg(rules, cursor, t_args, 1);
 	arg3 = get_arg(rules, cursor, t_args, 2);
 	if (t_args[0] == IND_CODE)
-		arg1 = get_value_from_battlefield(rules, cursor->position,
+		arg1 = (short)get_value_from_battlefield(rules, cursor->position,
 										  arg1 % IDX_MOD, REG_SIZE);
 	else if (t_args[0] == REG_CODE)
 		arg1 = cursor->reg[arg1];
 	if (t_args[1] == IND_CODE)
-		arg2 = get_value_from_battlefield(rules, cursor->position,
+		arg2 = (short)get_value_from_battlefield(rules, cursor->position,
 										  arg2 % IDX_MOD, REG_SIZE);
 	else if (t_args[1] == REG_CODE)
 		arg2 = cursor->reg[arg2];
@@ -109,7 +109,7 @@ void			ldi_func(t_rules *rules, t_champion *cursor, unsigned char *t_args)
 	arg2 = get_arg(rules, cursor, t_args, 1);
 	arg3 = get_arg(rules, cursor, t_args, 2);
 	if (t_args[0] == IND_CODE)
-		arg1 = get_value_from_battlefield(rules, cursor->position,
+		arg1 = (short)get_value_from_battlefield(rules, cursor->position,
 										  arg1 % IDX_MOD, REG_SIZE);
 	else if (t_args[0] == REG_CODE)
 		arg1 = cursor->reg[arg1];
